Updated to sending generally
WILL NOT COMPILE Bus.cpp awaiting howto info.

+++ Bus.h
@@ -7,10 +7,10 @@ class Bus {
       public:
 
       // Tell bus list of all outbound events we may send 
-      static void outboundEvents(uint16_t outbound[], uint16_t count);
+      static void outboundEvents(uint64_t outbound[], int16_t count);
       
       // Tell bus list of all sender:events that we are interested in
-      static void inboundEvents(uint64_t inbound[], uint16_t count);
+      static void inboundEvents(uint64_t inbound[], int16_t count);
 
       // Tell bus the callback function to invoke when a listened to event arrives.
       static void setCallback(EVENT_CALLBACK callback);
@@ -19,7 +19,7 @@ class Bus {
       static void ready();
 
       // Send an event on the bus
-      static void sendEvent(uint16_t eventid);
+      static void sendEvent(uint64_t eventid);
       
       // Arduino activity loop called repeatedly
       static void loop();
